requirements.txt
openai==1.3.5
python-telegram-bot==20.3
python-dotenv==1.0.0
